Move KtLintWorkAction and KtLintParameters into separate file.
This file has new package suffix - "worker".
