+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/WorkerApiKtLintRunner.kt
@@ -1,13 +1,8 @@
 package org.jlleitschuh.gradle.ktlint
 
-import org.gradle.api.file.ConfigurableFileCollection
 import org.gradle.api.file.FileCollection
-import org.gradle.api.file.RegularFileProperty
-import org.gradle.api.provider.Property
-import org.gradle.process.ExecOperations
-import org.gradle.workers.WorkAction
-import org.gradle.workers.WorkParameters
 import org.gradle.workers.WorkerExecutor
+import org.jlleitschuh.gradle.ktlint.worker.KtLintWorkAction
 import java.io.File
 import javax.inject.Inject
 
@@ -24,26 +19,6 @@ abstract class WorkerApiKtLintRunner : KtLintRunner {
             params.ktlintVersion.set(ktlintVersion)
             params.ignoreFailures.set(ignoreFailures)
             params.ktlintArgsFile.set(ktlintArgsFile)
-        }
-    }
-}
-
-interface KtLintParameters : WorkParameters {
-    val ktlintClasspath: ConfigurableFileCollection
-    val ktlintVersion: Property<String>
-    val ignoreFailures: Property<Boolean>
-    val ktlintArgsFile: RegularFileProperty
-}
-
-abstract class KtLintWorkAction : WorkAction<KtLintParameters> {
-    @get:Inject abstract val execOperations: ExecOperations
-
-    override fun execute() {
-        execOperations.javaexec {
-            it.classpath = parameters.ktlintClasspath
-            it.main = resolveMainClassName(parameters.ktlintVersion.get())
-            it.isIgnoreExitValue = parameters.ignoreFailures.get()
-            it.args("@${parameters.ktlintArgsFile.asFile.get()}")
         }
     }
 }
